+++ src/test/java/gov/usgs/owi/nldi/controllers/NetworkControllerFlowlineIT.java
@@ -43,7 +43,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUtTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13293474/navigation/UT/flowlines",
+				"/linked-data/comid/13293474/navigation/UT/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"7",
@@ -110,7 +110,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUtDiversionTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13294158/navigation/UT/flowlines",
+				"/linked-data/comid/13294158/navigation/UT/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"15",
@@ -124,7 +124,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUmTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13293474/navigation/UM/flowlines",
+				"/linked-data/comid/13293474/navigation/UM/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"4",
@@ -151,7 +151,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDmTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13296790/navigation/DM/flowlines",
+				"/linked-data/comid/13296790/navigation/DM/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"5",
@@ -164,7 +164,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDmDiversionsNotIncludedTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13294310/navigation/DM/flowlines",
+				"/linked-data/comid/13294310/navigation/DM/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"42",
@@ -191,7 +191,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDdTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13294310/navigation/DD/flowlines",
+				"/linked-data/comid/13294310/navigation/DD/flowlines?distance=9999",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"49",
@@ -218,9 +218,8 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	//PP Testing
 	@Test
 	public void getComidPpStopComidInvalidTest() throws Exception {
-
 		String actualbody = assertEntity(restTemplate,
-				"/linked-data/comid/13297246/navigation/PP/flowlines?stopComid=13297198",
+				"/linked-data/comid/13297246/navigation/PP/flowlines?distance=9999&stopComid=13297198",
 				HttpStatus.BAD_REQUEST.value(),
 				null,
 				null,
@@ -235,7 +234,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidPpStopComidTest() throws Exception {
 		assertEntity(restTemplate,
-				"/linked-data/comid/13297198/navigation/PP/flowlines?stopComid=13297246",
+				"/linked-data/comid/13297198/navigation/PP/flowlines?distance=9999&stopComid=13297246",
 				HttpStatus.OK.value(),
 				FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 				"12",
@@ -330,7 +329,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUtTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13293474/navigation/UT/flowlines",
+			"/linked-data/comid/13293474/navigation/UT/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"7",
@@ -397,7 +396,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUtDiversionTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13294158/navigation/UT/flowlines",
+			"/linked-data/comid/13294158/navigation/UT/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"15",
@@ -411,7 +410,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidUmTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13293474/navigation/UM/flowlines",
+			"/linked-data/comid/13293474/navigation/UM/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"4",
@@ -438,7 +437,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDmTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13296790/navigation/DM/flowlines",
+			"/linked-data/comid/13296790/navigation/DM/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"5",
@@ -451,7 +450,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDmDiversionsNotIncludedTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13294310/navigation/DM/flowlines",
+			"/linked-data/comid/13294310/navigation/DM/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"42",
@@ -478,7 +477,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidDdTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13294310/navigation/DD/flowlines",
+			"/linked-data/comid/13294310/navigation/DD/flowlines?distance=9999",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"49",
@@ -506,7 +505,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidPpStopComidInvalidTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13297246/navigation/PP/flowlines?stopComid=13297198",
+			"/linked-data/comid/13297246/navigation/PP/flowlines?distance=9999&stopComid=13297198",
 			HttpStatus.BAD_REQUEST.value(),
 			null,
 			null,
@@ -519,7 +518,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void getComidPpStopComidTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13297198/navigation/PP/flowlines?stopComid=13297246",
+			"/linked-data/comid/13297198/navigation/PP/flowlines?distance=9999&stopComid=13297246",
 			HttpStatus.OK.value(),
 			FlowLineTransformer.FLOW_LINES_COUNT_HEADER,
 			"12",
@@ -598,7 +597,7 @@ public class NetworkControllerFlowlineIT extends BaseIT {
 	@Test
 	public void badNavigationModeTestNavigation() throws Exception {
 		assertEntity(restTemplate,
-			"/linked-data/comid/13297198/navigation/XX/flowlines",
+			"/linked-data/comid/13297198/navigation/XX/flowlines?distance=9999",
 			HttpStatus.BAD_REQUEST.value(),
 			null,
 			null,